allow multiple copies of one ore
sieve drops are generated by parsing a JSON file. JSON does not support duplicate keys, so if an ore appears twice for the same block (i.e. at different mesh levels), the latter instance overwrites the former. This workaround allows for the JSON to contain multiple keys in different cases, which gsecore will parse as lowercase.